# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API for Oficina Mecânica Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /veiculos:
    get:
      operationId: listVeiculos
      tags: [veiculos]
      summary: List all vehicles
      parameters:
        - name: clienteId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Veiculo"
    post:
      operationId: createVeiculo
      tags: [veiculos]
      summary: Create a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVeiculo"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Veiculo"
  /veiculos/{id}:
    get:
      operationId: getVeiculo
      tags: [veiculos]
      summary: Get a vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Veiculo"
    put:
      operationId: updateVeiculo
      tags: [veiculos]
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVeiculo"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Veiculo"
    delete:
      operationId: deleteVeiculo
      tags: [veiculos]
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PECAS (ESTOQUE)
  /pecas:
    get:
      operationId: listPecas
      tags: [pecas]
      summary: List all parts
      responses:
        "200":
          description: List of parts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Peca"
    post:
      operationId: createPeca
      tags: [pecas]
      summary: Create a part
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePeca"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peca"
  /pecas/{id}:
    get:
      operationId: getPeca
      tags: [pecas]
      summary: Get a part by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peca"
    put:
      operationId: updatePeca
      tags: [pecas]
      summary: Update a part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePeca"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peca"
    delete:
      operationId: deletePeca
      tags: [pecas]
      summary: Delete a part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /pecas/{id}/movimentacao:
    post:
      operationId: movimentarPeca
      tags: [pecas]
      summary: Stock movement (entrada/saida)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovimentacaoPeca"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Peca"

  # SERVICOS (catalogo)
  /servicos:
    get:
      operationId: listServicos
      tags: [servicos]
      summary: List all services in catalog
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Servico"
    post:
      operationId: createServico
      tags: [servicos]
      summary: Create a service in catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServico"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Servico"
  /servicos/{id}:
    put:
      operationId: updateServico
      tags: [servicos]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServico"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Servico"
    delete:
      operationId: deleteServico
      tags: [servicos]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ORDENS DE SERVICO
...
components:
  schemas:
...
    Veiculo:
      type: object
      properties:
        id:
          type: integer
        clienteId:
          type: integer
        clienteNome:
          type: string
        placa:
          type: string
        modelo:
          type: string
        marca:
          type: string
        ano:
          type: integer
        km:
          type: integer
        cor:
          type: string
        observacoes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - clienteId
        - placa
        - modelo
        - marca
        - ano
        - createdAt
    CreateVeiculo:
      type: object
      properties:
        clienteId:
          type: integer
        placa:
          type: string
        modelo:
          type: string
        marca:
          type: string
        ano:
          type: integer
        km:
          type: integer
        cor:
          type: string
        observacoes:
          type: string
      required:
        - clienteId
        - placa
        - modelo
        - marca
        - ano
    Peca:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        codigo:
          type: string
        quantidade:
          type: integer
        quantidadeMinima:
          type: integer
        valorCusto:
          type: number
        valorVenda:
          type: number
        estoqueAlerta:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - nome
        - quantidade
        - quantidadeMinima
        - valorCusto
        - valorVenda
        - estoqueAlerta
        - createdAt
    CreatePeca:
      type: object
      properties:
        nome:
          type: string
        codigo:
          type: string
        quantidade:
          type: integer
        quantidadeMinima:
          type: integer
        valorCusto:
          type: number
        valorVenda:
          type: number
      required:
        - nome
        - quantidade
        - quantidadeMinima
        - valorCusto
        - valorVenda
    MovimentacaoPeca:
      type: object
      properties:
        tipo:
          type: string
          enum: [entrada, saida]
        quantidade:
          type: integer
        motivo:
          type: string
      required:
        - tipo
        - quantidade
    Servico:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: string
        valorPadrao:
          type: number
        createdAt:
          type: string
      required:
        - id
        - nome
        - valorPadrao
        - createdAt
    CreateServico:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        valorPadrao:
          type: number
      required:
        - nome
        - valorPadrao
...