# lib/api-zod/src/generated/types/createVeiculo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API for Oficina Mecânica Management System
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVeiculo {
  clienteId: number;
  placa: string;
  modelo: string;
  marca: string;
  ano: number;
  km?: number;
  cor?: string;
  observacoes?: string;
}